requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
django-bootstrap-v5==1.0.11
django-registration==3.2
django-registration-redux==2.10
python-decouple==3.6
soupsieve==2.3.1

